src/test/java/org/simpleapp/config/TokenInterceptorTest.java
package org.simpleapp.config;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.mock.web.MockHttpServletRequest;
import org.springframework.mock.web.MockHttpServletResponse;

import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
public class TokenInterceptorTest {

    @Value("${security.token}")
    private String expectedToken;

    @Autowired
    private TokenInterceptor tokenInterceptor;

    @Test
    public void testPreHandleValidToken() throws Exception {
        MockHttpServletRequest request = new MockHttpServletRequest();
        MockHttpServletResponse response = new MockHttpServletResponse();

        request.addHeader("Token", expectedToken);

        boolean result = tokenInterceptor.preHandle(request, response, new Object());

        assertTrue(result);
    }

    @Test
    public void testPreHandleInvalidToken() throws Exception {
        MockHttpServletRequest request = new MockHttpServletRequest();
        MockHttpServletResponse response = new MockHttpServletResponse();

        request.addHeader("Token", "invalid_token");

        boolean result = tokenInterceptor.preHandle(request, response, new Object());

        assertFalse(result);
        assertEquals(403, response.getStatus());
    }

    @Test
    public void testPreHandleMissingToken() throws Exception {
        MockHttpServletRequest request = new MockHttpServletRequest();
        MockHttpServletResponse response = new MockHttpServletResponse();

        boolean result = tokenInterceptor.preHandle(request, response, new Object());

        assertFalse(result);
        assertEquals(403, response.getStatus());
    }
}
